split test file
